+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2] - 2022-08-17
+### Added
+- Dataplex Assets Module
+
+### Changed
+- Bumped minimum version of the provider to 4.32.0
+
 ## [0.1] - 2022-08-05
 ### Added
 - Initial Release